requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.20
pillow==11.3.0
Jinja2==3.1.6
numpy==1.26.4
tensorflow-macos==2.13.0